# packages/functions/api/openapi.yaml
openapi: 3.0.3
info:
  title: Meal Planner API
  description: API for managing meal planning schedules, groups, and members
  version: 1.0.0
servers:
  - url: https://api-meal-planner.isnan.eu
    description: Production server
paths:
...
  /api/groups/{groupId}/invites:
    post:
      summary: Create an invite
      description: Generate a reusable invite link for the group. Only the GroupAdmin may call this endpoint. The invite expires after 7 days.
      operationId: createInvite
      tags:
        - Invites
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the group
      responses:
        '201':
          description: Invite created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
        '401':
          description: Unauthorized
        '403':
          description: Caller is not the GroupAdmin
        '404':
          description: Group not found
    get:
      summary: List active invites
      description: List all active (non-expired) invites for the group. Only the GroupAdmin may call this endpoint.
      operationId: listInvites
      tags:
        - Invites
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the group
      responses:
        '200':
          description: Invites listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListInvitesResponse'
        '401':
          description: Unauthorized
        '403':
          description: Caller is not the GroupAdmin
...
  /api/invites/{code}:
    get:
      summary: Get invite details
      description: Retrieve details about an invite (group name and expiry). Returns 404 for missing, expired, or revoked invites — these cases are intentionally indistinguishable to the caller.
      operationId: getInvite
      tags:
        - Invites
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          description: The invite code
      responses:
        '200':
          description: Invite details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteDetailsResponse'
        '401':
          description: Unauthorized
        '404':
          description: Invite not found, expired, or revoked
  /api/invites/{code}/redeem:
    post:
      summary: Redeem an invite
      description: "Redeem an invite code to join the associated group as a regular member. This operation is idempotent — calling it when already a member returns a 200 with alreadyMember=true rather than an error."
      operationId: redeemInvite
      tags:
        - Invites
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          description: The invite code to redeem
      responses:
        '200':
          description: Invite redeemed (or caller was already a member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedeemInviteResponse'
        '401':
          description: Unauthorized
        '404':
          description: Invite not found, expired, or revoked
...
components:
  securitySchemes:
    cognitoAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Cognito User Pool JWT token
  schemas:
...
    InviteResponse:
      type: object
      properties:
        code:
          type: string
          description: Unique invite code (alphanumeric, 8 characters)
        expiresAt:
          type: string
          format: date-time
          description: Timestamp when the invite expires (7 days from creation)
    ListInvitesResponse:
      type: object
      properties:
        invites:
          type: array
          items:
            $ref: '#/components/schemas/InviteResponse'
          description: List of active (non-expired) invites for the group
    InviteDetailsResponse:
      type: object
      properties:
        groupName:
          type: string
          description: Name of the group the invite grants access to
        expiresAt:
          type: string
          format: date-time
          description: Timestamp when the invite expires
    RedeemInviteResponse:
      type: object
      properties:
        groupId:
          type: string
          description: Unique identifier of the group the caller joined
        groupName:
          type: string
          description: Name of the group the caller joined
        alreadyMember:
          type: boolean
          description: True if the caller was already a member of the group before redemption
...